+++ src/routes/index.js
@@ -4,13 +4,12 @@ const express = require('express');
 const router = express.Router();
 
 router.get('/', (req, res) => {
-    res.render('index');
+  res.render('index');
 });
 
 router.get('/about', (req, res) => {
-    res.render('about');
+  res.render('about');
 });
-
 
 
 module.exports = router;